Continued refactor
