+++ screenflow/font_manager.py
@@ -1,99 +1,26 @@
 #!/usr/bin/python
 
-""" Module that provides FontManager base class. """
-
-import logging
-import pygame
-
-from constants import BLACK
-
-# Configure logger.
-logging.basicConfig()
-logger = logging.getLogger(__name__)
-
-
-class FontHolder(object):
-    """Holder class for a font and a text color.
-
-    When accessing a font through getter method, if font has not been settled,
-    default system Arial font is used.
-    """
-
-    def __init__(self, default_size):
-        """ Default constructor. """
-        self._font = None
-        self.text_color = BLACK
-        self.default_size = default_size
-
-    @property
-    def font(self):
-        """Delegate font instance getter. Creating
-        one if not available using arial system font.
-
-        :returns: Font instance to use.
-        """
-        if self._font is None:
-            logging.debug('Font not settled, using default Arial from system')
-            self._font = pygame.font.SysFont('arial', self.default_size)
-        return self._font
-
-    @font.setter
-    def font(self, font):
-        """Setter for font value.
-
-        :param font: Font value to set.
-        """
-        self._font = font
-
-    def __call__(self, text):
-        """Function call overloading that allows holder
-        to be used as a text sizer.
-
-        :param text: Text to evaluate rendering for.
-        :returns: Rendering size of the given text.
-        """
-        return self.font.size(text)
-
-
-def draw_text(text, surface, holder, position):
-    """Simple methods that draws text on the given surface
-    using the given font holder for resources.
-
-    :param text: Text to draw.
-    :param surface: Surface to draw text into.
-    :param holder: FontHolder instance to use for getting font and text color.
-    :param position: Position of the text to draw relative to the surface.
-    """
-    surface.blit(holder.font.render(text, 1, holder.text_color, None), position)
+"""
+"""
 
 
 class FontManager(object):
-    """A FontManager is responsible for storing two fonts :
+    """ FontManager is a simple font caching factory. """
 
-    - Primary font that will be used for important text display.
-    - Secondary font that will be used for tips, subheading, and so on.
-
-    """
-
-    def __init__(self):
+    def __init__(self, font_factory):
         """ Default constructor. """
-        self.primary = FontHolder(15)
-        self.secondary = FontHolder(10)
+        self.fonts = {}
+        self.font_factory = font_factory
 
-    def draw_primary_text(self, text, surface, position):
-        """Draws text to screen using primary font.
+    def get(self, name, size):
+        """Font access method. Creates the font instance if not exists.
 
-        :param text: Text to draw.
-        :param surface: Surface to draw text into.
-        :param position: Position of the text to draw relative to the surface.
+        :param name: Name of the font to get.
+        :param size: Size of the font to get.
+        :returns: Required font instance.
         """
-        draw_text(text, surface, self.primary, position)
-
-    def draw_secondary_text(self, text, surface, position):
-        """Draws text to screen using secondary font.
-
-        :param text: Text to draw.
-        :param surface: Surface to draw text into.
-        :param position: Position of the text to draw relative to the surface.
-        """
-        draw_text(text, surface, self.secondary, position)
+        if name not in self.fonts.keys():
+            self.fonts[name] = {}
+        if size not in self.fonts.keys():
+            self.fonts[name][size] = font_factory(name, size)
+        return self.fonts[name][size]
